+++ src/pages/careers/_be_square_banner.tsx
