+++ src/pages/careers/_layout-components/_career_container.ts
