+++ src/pages/careers/locations/_location-layout.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
-import CareerContainer from '../_layout-components/CareerContainer'
+import CareerContainer from '../_layout-components/_career_container'
 import { locationsTypes } from '../_model/_locations/_locations.types'
 import device from 'themes/device'
 import { SectionContainer, Flex } from 'components/containers'
